requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
opencv-python==4.8.0.76
numpy==1.25.2
Pillow==10.0.0
